+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This release contains multiple **breaking changes** in the Go API. It is suppose
 ### Added
 
 - `Tf.Cmd` also sets `Stdin` to `os.Stdin`.
+- Add `Taskflow.Tasks` and `Taskflow.Params` to allow introspection of the registered tasks and parameters.
 
 ### Changed